Cleaning up src/modules/modules.py into several files.
