+++ src/modules/edge_probing.py
@@ -10,7 +10,7 @@ from allennlp.modules.span_extractors import EndpointSpanExtractor, SelfAttentiv
 from torch.autograd import Variable
 
 from ..tasks.edge_probing import EdgeProbingTask
-from . import modules
+from .simple_modules import Classifier
 
 
 class EdgeClassifierModule(nn.Module):
@@ -99,7 +99,7 @@ class EdgeClassifierModule(nn.Module):
         clf_input_dim = self.span_extractors[1].get_output_dim()
         if not self.single_sided:
             clf_input_dim += self.span_extractors[2].get_output_dim()
-        self.classifier = modules.Classifier.from_params(clf_input_dim, task.n_classes, task_params)
+        self.classifier = Classifier.from_params(clf_input_dim, task.n_classes, task_params)
 
     def forward(
         self,
