+++ src/modules/seq2seq_decoder.py
@@ -20,7 +20,7 @@ from overrides import overrides
 from torch.nn.modules.linear import Linear
 from torch.nn.modules.rnn import LSTMCell
 
-from .modules import Pooler
+from .simple_modules import Pooler
 
 
 class Seq2SeqDecoder(Model):
