+++ tests/test_models.py
@@ -1,0 +1,6 @@
+import unittest
+
+
+class TestModel(unittest.TestCase):
+    def test_import(self):
+        from src.models import build_model
